+++ doc/source/conf.py
@@ -24,7 +24,7 @@ copyright = '2020-2021, Geoffroy Lesur et al.'
 author = 'Geoffroy Lesur'
 
 # The full version, including alpha/beta/rc tags
-release = '0.1'
+release = '0.7-dev'
 
 
 # -- General configuration ---------------------------------------------------