+++ src/dataBlock/evolveStage.cpp
@@ -12,7 +12,7 @@
 void DataBlock::EvolveStage() {
   idfx::pushRegion("DataBlock::EvolveStage");
   // Compute current when needed
-  if(hydro.needCurrent) hydro.CalcCurrent();
+  if(hydro.needExplicitCurrent) hydro.CalcCurrent();
 
   // Loop on all of the directions
   for(int dir = 0 ; dir < DIMENSIONS ; dir++) {
@@ -23,7 +23,7 @@ void DataBlock::EvolveStage() {
     hydro.CalcRiemannFlux(dir, this->t);
 
     // Step 2.5: compute intercell parabolic flux when needed
-    if(hydro.haveParabolicTerms) hydro.CalcParabolicFlux(dir, this->t);
+    if(hydro.haveExplicitParabolicTerms) hydro.CalcParabolicFlux(dir, this->t);
 
     // Step 3: compute the resulting evolution of the conserved variables, stored in Uc
     hydro.CalcRightHandSide(dir, this->t, this->dt);
@@ -35,9 +35,10 @@ void DataBlock::EvolveStage() {
 #if MHD == YES && DIMENSIONS >= 2
   // Compute the field evolution according to CT
   hydro.emf.CalcCornerEMF(this->t);
-  if(hydro.haveResistivity || hydro.haveAmbipolar) hydro.emf.CalcNonidealEMF(this->t);
+  if(hydro.resistivityStatus.isExplicit || hydro.ambipolarStatus.isExplicit)
+    hydro.emf.CalcNonidealEMF(this->t);
   hydro.emf.EnforceEMFBoundary();
-  hydro.emf.EvolveMagField(this->t, this->dt);
+  hydro.emf.EvolveMagField(this->t, this->dt, hydro.Vs);
   hydro.ReconstructVcField(hydro.Uc);
 #endif